% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtm.R
\name{extract_keywords}
\alias{extract_keywords}
\title{Extract top TF-IDF keywords}
\usage{
extract_keywords(x, n = 1L, already_tfidf = FALSE)
}
\arguments{
\item{x}{A document-term matrix or TF-IDF matrix.}

\item{n}{Number of keywords per document.}

\item{already_tfidf}{Logical; is \code{x} already weighted?}
}
\value{
A data frame with document, rank, term, and weight.
}
\description{
Extract top TF-IDF keywords
}
