% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{f_measure}
\alias{f_measure}
\title{F-measure}
\usage{
f_measure(precision, recall, beta = 1)
}
\arguments{
\item{precision}{Numeric precision.}

\item{recall}{Numeric recall.}

\item{beta}{Relative weight assigned to recall.}
}
\value{
Numeric F-measure.
}
\description{
F-measure
}
