% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{read_label_file}
\alias{read_label_file}
\title{Read a one-label-per-line file}
\usage{
read_label_file(path, type = c("character", "integer", "numeric", "factor"))
}
\arguments{
\item{path}{Path to a text file.}

\item{type}{Return labels as character, integer, numeric, or factor.}
}
\value{
A vector of labels.
}
\description{
Read a one-label-per-line file
}
